TST move tests to new file and fixes from code review
